+++ lib/ui/groups/GroupMemberNotificationsPanel.dart
@@ -119,7 +119,7 @@ class _GroupMemberNotificationsPanelState extends State<GroupMemberNotifications
         enabled: _toggleButtonEnabled,
         borderRadius: BorderRadius.zero,
         label: Localization().getStringEx(
-            "panel.group_member_notifications.override_notifications.label", "Override my notification preferences for this group"),
+            "panel.group_member_notifications.override_notifications.label", "Use these notification preferences for this group"),
         toggled: memberPreferences?.overridePreferences ?? false,
         onTap: _toggleButtonEnabled ? _onToggleOverrideNotificationPreferences : null,
         textStyle: _toggleButtonEnabled
@@ -175,7 +175,6 @@ class _GroupMemberNotificationsPanelState extends State<GroupMemberNotifications
   }
 
   _buildGlobalNotificationsInfo(){
-    // bool groupNotificationsEnabled = FirebaseMessaging().notifyGroupUpdates == true;
     bool groupPostNotificationsEnabled = FirebaseMessaging().notifyGroupPostUpdates == true;
     bool groupEventsNotificationsEnabled = FirebaseMessaging().notifyGroupEventsUpdates == true;
     bool groupInvitationsNotificationsEnabled = FirebaseMessaging().notifyGroupInvitationsUpdates == true;
@@ -187,40 +186,30 @@ class _GroupMemberNotificationsPanelState extends State<GroupMemberNotifications
           child: Container(
               padding: EdgeInsets.symmetric(horizontal: 10, vertical: 10),
               color: Colors.white,
-              child: Text(Localization().getStringEx('panel.group_member_notifications.member.info.heading.msg', 'Global Group Notification Settings'), //TBD localize
+              child: Text(Localization().getStringEx('panel.group_member_notifications.member.info.heading.msg', 'Current group notification preferences'), //TBD localize
                   style: Styles().textStyles?.getTextStyle("widget.title.regular")),
           )
         )
     ],));
 
-    // widgets.add(ToggleRibbonButton(
-    //     borderRadius: BorderRadius.zero,
-    //     label: Localization().getStringEx("panel.settings.notifications.group_updates", "Group Updates"),
-    //     toggled: groupNotificationsEnabled,
-    //     textStyle: groupNotificationsEnabled ? Styles().textStyles?.getTextStyle("panel.group_member_notifications.toggle_button.title.small.enabled") : Styles().textStyles?.getTextStyle("panel.settings.toggle_button.title.fat.disabled")
-    // ));
     widgets.add(Row(children: [Expanded(child: Container(color: Styles().colors!.white, child: Padding(padding: EdgeInsets.only(left: 10), child: Column(children: [
-      ToggleRibbonButton(
+      _DisabledToggleButton(
           toggled: groupPostNotificationsEnabled,
-          borderRadius: BorderRadius.zero,
           label: Localization().getStringEx("panel.settings.notifications.group_updates.posts.label", "Posts"),
           textStyle: groupPostNotificationsEnabled ? Styles().textStyles?.getTextStyle("panel.group_member_notifications.toggle_button.title.small.enabled"): Styles().textStyles?.getTextStyle("panel.settings.toggle_button.title.small.disabled")
       ),
-      ToggleRibbonButton(
+      _DisabledToggleButton(
           toggled: groupEventsNotificationsEnabled,
-          borderRadius: BorderRadius.zero,
           label: Localization().getStringEx("panel.settings.notifications.group_updates.event.label", "Event"),
           textStyle: groupEventsNotificationsEnabled ? Styles().textStyles?.getTextStyle("panel.group_member_notifications.toggle_button.title.small.enabled"): Styles().textStyles?.getTextStyle("panel.settings.toggle_button.title.small.disabled")
       ),
-      ToggleRibbonButton(
+      _DisabledToggleButton(
           toggled: groupInvitationsNotificationsEnabled,
-          borderRadius: BorderRadius.zero,
           label: Localization().getStringEx("panel.settings.notifications.group_updates.invitations.label", "Invitations"),
           textStyle: groupInvitationsNotificationsEnabled ? Styles().textStyles?.getTextStyle("panel.group_member_notifications.toggle_button.title.small.enabled"): Styles().textStyles?.getTextStyle("panel.settings.toggle_button.title.small.disabled")
       ),
-      ToggleRibbonButton(
+      _DisabledToggleButton(
           toggled:  groupPollsNotificationsEnabled,
-          borderRadius: BorderRadius.zero,
           label: Localization().getStringEx("panel.settings.notifications.group_updates.polls.label", "Polls"),
           textStyle: groupPollsNotificationsEnabled ? Styles().textStyles?.getTextStyle("panel.group_member_notifications.toggle_button.title.small.enabled"): Styles().textStyles?.getTextStyle("panel.settings.toggle_button.title.small.disabled")
       )
@@ -374,6 +363,19 @@ class _GroupMemberNotificationsPanelState extends State<GroupMemberNotifications
       setStateIfMounted(() {});
     }
   }
+}
+
+class _DisabledToggleButton extends ToggleRibbonButton{
+  static Map<bool, Widget> _rightIcons= {
+    true: Styles().images?.getImage("images/green-check-mark.png", excludeFromSemantics: true, size: 20) ?? Container(),
+    false: Styles().images?.getImage("images/icon-x-orange.png", excludeFromSemantics: true, size: 20) ?? Container(),
+  };
+
+  _DisabledToggleButton( {String? label,
+      bool? toggled,
+      void Function()? onTap,
+      TextStyle? textStyle,})
+      : super(label: label, toggled: (toggled == true), onTap: onTap, textStyle: textStyle, padding: const EdgeInsets.symmetric(horizontal: 16, vertical: 8),rightIcons: _rightIcons);
 }
 
 class _EnabledToggleButton extends ToggleRibbonButton {