+++ CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+### Changed
+- Group Notification Settings panel: update styling [#2538](https://github.com/rokwire/illinois-app/issues/2538).
+
 ## [4.2.25] - 2022-11-23
 ### Changed
 - Research Project updates [#2563](https://github.com/rokwire/illinois-app/issues/2563).